Move regexes module to core/regexes
