+++ pyanaconda/core/regexes.py
