+++ pyanaconda/input_checking.py
@@ -22,9 +22,8 @@ import pwquality
 
 from pyanaconda.isignal import Signal
 from pyanaconda.core.i18n import _
-from pyanaconda.core import constants
+from pyanaconda.core import constants, regexes
 from pyanaconda import users
-from pyanaconda import regexes
 
 from pyanaconda.anaconda_loggers import get_module_logger
 log = get_module_logger(__name__)
