+++ pyanaconda/ui/gui/spokes/advstorage/dasd.py
@@ -25,7 +25,7 @@ from gi.repository import BlockDev as blockdev
 from pyanaconda.ui.gui import GUIObject
 from pyanaconda.core.async_utils import async_action_nowait
 from pyanaconda.storage_utils import try_populate_devicetree
-from pyanaconda.regexes import DASD_DEVICE_NUMBER
+from pyanaconda.core.regexes import DASD_DEVICE_NUMBER
 from pyanaconda.threading import threadMgr, AnacondaThread
 from pyanaconda.core.timer import Timer
 from pyanaconda.core import constants
