+++ pyanaconda/ui/gui/spokes/advstorage/zfcp.py
@@ -27,7 +27,7 @@ from pyanaconda.ui.gui import GUIObject
 from pyanaconda.core.async_utils import async_action_nowait
 from pyanaconda.core.timer import Timer
 from pyanaconda.storage_utils import try_populate_devicetree
-from pyanaconda.regexes import DASD_DEVICE_NUMBER, ZFCP_WWPN_NUMBER, ZFCP_LUN_NUMBER
+from pyanaconda.core.regexes import DASD_DEVICE_NUMBER, ZFCP_WWPN_NUMBER, ZFCP_LUN_NUMBER
 from pyanaconda.threading import threadMgr, AnacondaThread
 from pyanaconda.core import constants
