+++ pyanaconda/users.py
@@ -25,7 +25,7 @@ from contextlib import contextmanager
 from pyanaconda.core import iutil
 from pyanaconda.core.iutil import strip_accents
 from pyanaconda.errors import errorHandler, PasswordCryptError, ERROR_RAISE
-from pyanaconda.regexes import GROUPLIST_FANCY_PARSE, NAME_VALID, PORTABLE_FS_CHARS, GROUPLIST_SIMPLE_VALID
+from pyanaconda.core.regexes import GROUPLIST_FANCY_PARSE, NAME_VALID, PORTABLE_FS_CHARS, GROUPLIST_SIMPLE_VALID
 import crypt
 from pyanaconda.core.i18n import _
 import re
