+++ tests/regex_tests/dasd_name_test.py
@@ -21,7 +21,7 @@
 import unittest
 
 from regexcheck import regex_match
-from pyanaconda.regexes import DASD_DEVICE_NUMBER
+from pyanaconda.core.regexes import DASD_DEVICE_NUMBER
 
 
 class DASDNameRegexTestCase(unittest.TestCase):
