+++ tests/regex_tests/ibft_device_name_test.py
@@ -21,7 +21,7 @@
 #
 import unittest
 
-from pyanaconda.regexes import IBFT_CONFIGURED_DEVICE_NAME
+from pyanaconda.core.regexes import IBFT_CONFIGURED_DEVICE_NAME
 
 def _run_tests(testcase, expression, goodlist, badlist):
     got_error = False
