+++ tests/regex_tests/iscsi_name_test.py
@@ -17,7 +17,7 @@
 import unittest
 
 from regexcheck import regex_match
-from pyanaconda.regexes import ISCSI_IQN_NAME_REGEX, ISCSI_EUI_NAME_REGEX
+from pyanaconda.core.regexes import ISCSI_IQN_NAME_REGEX, ISCSI_EUI_NAME_REGEX
 
 class iSCSIiqnnameRegexTestCase(unittest.TestCase):
     def iqnname_test(self):
