+++ tests/regex_tests/zfcp_name_test.py
@@ -21,7 +21,7 @@
 import unittest
 
 from regexcheck import regex_match
-from pyanaconda.regexes import ZFCP_LUN_NUMBER, ZFCP_WWPN_NUMBER
+from pyanaconda.core.regexes import ZFCP_LUN_NUMBER, ZFCP_WWPN_NUMBER
 
 
 class ZFCPNameRegexTestCase(unittest.TestCase):
